requirements.txt
requests==2.23.0
python_telegram_bot==12.6.1
flake8==3.8.3
coverage==5.2.1
typer==0.3.2
gunicorn == 20.0.4
